requirements.txt
cached-property==1.5.1    # via django-url-filter
django-url-filter==0.3.12
django==2.1.7
djangorestframework-jsonapi==2.7.0
djangorestframework==3.9.1
enum-compat==0.0.2        # via django-url-filter
inflection==0.3.1         # via djangorestframework-jsonapi
psycopg2-binary==2.7.7
python-decouple==3.1
pytz==2018.9              # via django
six==1.12.0               # via django-url-filter, djangorestframework-jsonapi
